Move getVersionType to its proper file location
